Organize the file layout
