+++ src/Polygon_Intersection/algorithms/Bentley_Ottmann/AVLtree.py
