+++ src/Polygon_Intersection/algorithms/polygon_orientation.py
